remove bindgen from proj
